Put PayloadServiceBuilder in generator
